Move expo module into different java/android package than app, to avoid build failure from eas build --local. Wasn't causing a problem for github action build or remote EAS build...
