+++ modules/expo-experiments/android/build.gradle
@@ -1,6 +1,6 @@
 apply plugin: 'com.android.library'
 
-group = 'uk.me.jeremygreen.expoexperiments'
+group = 'uk.me.jeremygreen.expoexperiments.module'
 version = '0.6.3'
 
 def expoModulesCorePlugin = new File(project(":expo-modules-core").projectDir.absolutePath, "ExpoModulesCorePlugin.gradle")
@@ -32,7 +32,7 @@ if (useManagedAndroidSdkVersions) {
 }
 
 android {
-  namespace "uk.me.jeremygreen.expoexperiments"
+  namespace "uk.me.jeremygreen.expoexperiments.module"
   defaultConfig {
     versionCode 1
     versionName "0.6.3"
